build fix

+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Game/CarlaEngine.cpp
@@ -30,6 +30,8 @@
 #include <carla/streaming/Server.h>
 #include <compiler/enable-ue4-macros.h>
 
+#include <cstdint>
+#include <sys/types.h>
 #include <thread>
 
 // =============================================================================
@@ -149,7 +151,7 @@ void FCarlaEngine::NotifyInitGame(const UCarlaSettings &Settings)
             auto sensor_id = *(reinterpret_cast<carla::streaming::detail::stream_id_type *>(Data.data()));
             // query dispatcher
             carla::streaming::detail::token_type token(Server.GetStreamingServer().GetToken(sensor_id));
-            carla::Buffer buf(reinterpret_cast<unsigned char *>(&token), (size_t) sizeof(token));
+            carla::Buffer buf(reinterpret_cast<unsigned char *>(&token), (uint32_t) sizeof(token));
             carla::log_info("responding with a token for port ", token.get_port());
             Secondary->Write(std::move(buf));
             break;
@@ -157,7 +159,7 @@ void FCarlaEngine::NotifyInitGame(const UCarlaSettings &Settings)
           case carla::multigpu::MultiGPUCommand::YOU_ALIVE:
           {
             std::string msg("Yes, I'm alive");
-            carla::Buffer buf((unsigned char *) msg.c_str(), (size_t) msg.size());
+            carla::Buffer buf((unsigned char *) msg.c_str(), (uint32_t) msg.size());
             carla::log_info("responding is alive command");
             Secondary->Write(std::move(buf));
             break;
@@ -170,7 +172,7 @@ void FCarlaEngine::NotifyInitGame(const UCarlaSettings &Settings)
             Server.GetStreamingServer().EnableForROS(sensor_id);
             // return a 'true'
             bool res = true;
-            carla::Buffer buf(reinterpret_cast<unsigned char *>(&res), (size_t) sizeof(bool));
+            carla::Buffer buf(reinterpret_cast<unsigned char *>(&res), (uint32_t) sizeof(bool));
             carla::log_info("responding ENABLE_ROS with a true");
             Secondary->Write(std::move(buf));
             break;
@@ -183,7 +185,7 @@ void FCarlaEngine::NotifyInitGame(const UCarlaSettings &Settings)
             Server.GetStreamingServer().DisableForROS(sensor_id);
             // return a 'true'
             bool res = true;
-            carla::Buffer buf(reinterpret_cast<unsigned char *>(&res), (size_t) sizeof(bool));
+            carla::Buffer buf(reinterpret_cast<unsigned char *>(&res), (uint32_t) sizeof(bool));
             carla::log_info("responding DISABLE_ROS with a true");
             Secondary->Write(std::move(buf));
             break;
@@ -194,7 +196,7 @@ void FCarlaEngine::NotifyInitGame(const UCarlaSettings &Settings)
             auto sensor_id = *(reinterpret_cast<carla::streaming::detail::stream_id_type *>(Data.data()));
             // query dispatcher
             bool res = Server.GetStreamingServer().IsEnabledForROS(sensor_id);
-            carla::Buffer buf(reinterpret_cast<unsigned char *>(&res), (size_t) sizeof(bool));
+            carla::Buffer buf(reinterpret_cast<unsigned char *>(&res), (uint32_t) sizeof(bool));
             carla::log_info("responding IS_ENABLED_ROS with: ", res);
             Secondary->Write(std::move(buf));
             break;
@@ -348,7 +350,7 @@ void FCarlaEngine::OnPostTick(UWorld *World, ELevelTick TickType, float DeltaSec
 
         // send frame data to secondary
         std::string Tmp(OutStream.str());
-        SecondaryServer->GetCommander().SendFrameData(carla::Buffer(std::move((unsigned char *) Tmp.c_str()), (size_t) Tmp.size()));
+        SecondaryServer->GetCommander().SendFrameData(carla::Buffer(std::move((unsigned char *) Tmp.c_str()), (uint32_t) Tmp.size()));
 
         GetCurrentEpisode()->GetFrameData().Clear();
       }
